ER-98 added short names, UI fixes and version++

+++ app/Erkc/app/src/main/kotlin/com/software/ssp/erkc/modules/history/valuehistory/ValueHistoryViewModel.kt
@@ -6,5 +6,5 @@ import com.software.ssp.erkc.data.realm.models.RealmIpuValue
 class ValueHistoryViewModel(
         val values: List<RealmIpuValue>,
         val total: Long,
-        val average: Long
+        val average: Double
 )
